requirements.txt
osmnx>=1.6.0
networkx>=3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
folium>=0.14.0
tqdm>=4.65.0
srtm.py>=0.3.7
rasterio>=1.3.0
matplotlib>=3.7.0
pytest>=7.4.0
pytest-cov>=4.1.0
streamlit>=1.28.0
streamlit-folium>=0.15.0